refactor: extracted countdown logic into useDashboardCompetition hook.
